Make functions that use VAD object part of the VAD object.
